ti-j72xx: add GPU feature support on J7 platform
Issue: LINUXEXEC-11606

At first, we add a new MACHINE_FEATURES option "gpu" to offer an alternative
so that customers can determine whether they would include GPU support.

The package mesa in oe-core layer provides libgles1/libgles2/egl/libgl/libgbm libraries.
But TI provides GPU specific libraries, so we need to disable mesa package and use
package "ti-img-rogue-umlibs" to provide libgles1/libgles2/egl/libgbm and package
"mesa-gl" to provide libgl.

The modifications as below:
1) adjust PREFERRED_PROVIDER for OpenGL libraries in onf/machine/ti-j72xx.conf.
2) adjust "COMPATIBLE_MACHINE" for package img-pvr-sdk/ti-img-rogue-driver/ti-img-rogue-umlibs.
3) in package ti-img-rogue-umlibs, change libraries install path in do_install since {libdir}
is "/usr/lib64" but not "/usr/lib" when multilib is enabled. And conserve pvr_dri.so in a fixed
path "/usr/lib/dir".
4) delete "glx" from PACKAGECONFIG of xserver-xorg since mesa which provides glx feature
is removed and ti-img-rogue-umlibs doesn't provide glx libraries and headers. And "xwayland"
is based on "glx", so it should also be removed.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>