ti-j72xx: add user space package devil
Issue: LINCD-4537

When running the vision demo "run_app_stereo.sh", it shows to need the
library "libIL.so.1", which is involved by package "devil" (Developer's Image Library).

The package "devil" is based on the GPU feature, so put in the
recipes-graphics directory and also no local downloadable mirror image.

For the package devil, directly pick the .bb from TI SDK, retain the TI
SDK patch "Remove-ILUT-dependency.patch" as-is and newly add one patch
"0001-DevIL-replace-lib-with-lib64-to-CMakeLists.txt.patch" to fit the
/usr/lib64 directory in our product.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>